refactor: UIResponder의 extension부분을 Util로 옮김
